Refactor Markets to call filterMarkets from helpers
